Can we simplify language and purpose here?
The goal of the guidance is commendable, but it tends to be a hard read, and I'm not sure it has to be.  First suggested edited pp here may not meet full intent, but since were really talking about saving money, reducing inefficiencies and waste while increasing transparency I'm wondering if something like this might work instead.  

Finally, the multiple in-line references to other GAO documents seems to overly complicate the text, and since you can embed hyperlinks, perhaps they don't need to be spelled out quite as much.  If the document is intended at all for broad consumption I'm wondering if a few edits in this light might go a long way towards improving the document and the clarity of the message.

For what it's worth -